Flyttet "Deres sykefraværstatistikk sammenlignet med bransjen" inn i egen komponent. Skjuler hele komponenten dersom det mangler data.
