--- src/Aktiviteter/Sykefraværsstatistikk.module.scss
@@ -1,5 +1,0 @@
-.statistikkContainer {
-    display: flex;
-    flex-wrap: wrap;
-    gap: 1rem;
-}
